Add dependency on ca-certificates to fix build on OBS
Commit 8c611a0fcd57 "Fix TLS connections by explicitly setting CERT_LOCATION"
tried to unbreak HTTPS connections by forcing the location of
`/etc/ssl/certs/ca-certificates.crt`, but it made build fail on OBS
because the buildsystem actually tries to check the existence of the
file at build time. Unlike on the Debian buildds, such file is not
available by default in the build chroots created by OBS since
`ca-certificates` is not in the pre-installed set.

Add the build-dependency on `ca-certificates` to ensure it is available
during the build on OBS.

The Debian upstream packaging also added the build-dep with the commit
from Sudip Mukherjee <sudipm.mukherjee@gmail.com>:

Add dependency on ca-certificates and libpcre3-dev

https://salsa.debian.org/debian/libgit2/-/commit/86ca7244128f1df7d1b24

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>